+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/watcher/FlinkAppHttpWatcher.java
@@ -186,7 +186,7 @@ public class FlinkAppHttpWatcher {
   public void doStop() {
     log.info(
         "FlinkAppHttpWatcher StreamPark Console will be shutdown,persistent application to database.");
-    WATCHING_APPS.forEach((k, v) -> applicationInfoService.persistMetrics(v));
+    WATCHING_APPS.forEach((k, v) -> applicationManageService.persistMetrics(v));
   }
 
   /**
@@ -515,7 +515,7 @@ public class FlinkAppHttpWatcher {
     } else {
       WATCHING_APPS.put(application.getId(), application);
     }
-    applicationInfoService.persistMetrics(application);
+    applicationManageService.persistMetrics(application);
   }
 
   /**
